ci: add missing pip-compile lockfile to pip-compile-dev workflow
The lockfiles for the pip-compile session itself were never updated
because I forgot to add them to the update workflow.

The pr_labeler lockfile is still excluded, but that's intentional for
now. I'd like this to happen manually or as part of a separate automated
workflow for now, as there's no automated tests for that script as of
yet, so I don't want it to randomly break when we merge these general
pip-compile-dev PRs.